fix warnings/errors

+++ JSQMessagesViewController/Controllers/JSQMessagesViewController.m
@@ -513,7 +513,7 @@ static void JSQInstallWorkaroundForSheetPresentationIssue26295020(void) {
     }
 
     JSQMessagesCollectionViewCell *cell = [collectionView dequeueReusableCellWithReuseIdentifier:cellIdentifier forIndexPath:indexPath];
-    cell.accessibilityIdentifier = [NSString stringWithFormat:@"(%ld, %ld)", indexPath.section, indexPath.row];
+    cell.accessibilityIdentifier = [NSString stringWithFormat:@"(%ld, %ld)", (long)indexPath.section, (long)indexPath.row];
     cell.delegate = collectionView;
 
     if (!isMediaMessage) {
